+++ CHANGES.rst
@@ -4,6 +4,10 @@ Changelog
 3.1 - Unreleased
 ----------------
 
+- Implement Python 3.6 ``readline.set_auto_history`` and the corresponding
+  ``history.auto``.
+  [stefan]
+
 - Replace deprecated ``python setup.py test`` in tox.ini.
   [stefan]